migrated to MDEditor editor component
This required removing old components, altering existing tests, etc.

The reason for this is that Draft-js is far too complicated and messy to work with. Markdown is far easier to work with (no kidding)

Added some styling for the 'post' card images so they stay the same width and height no matter the underlying image size.