configured a temporary cookie verification setup
uses `useEffect` and `js-cookie` to verify that a cookie exists with a valid name.

This isn't a long term solution since a valid JWT isn't being verified. It's just for UI purposes

Disabled the cookie re-route logic until it can be more permanently implemented so that tests don't break.

Also created an '/admin' route to create a cookie in the browser via the UI.

Added an alert to the SignUp component to warn the user that they need to sign up for an account to view the home page