wired up the verification route changes in the backend to the frontend
added two new routes to the router, `/verify-email` and `/verify/:token`.

`/verify-email` is a simple prompt that lets the user know that a verification email has been sent to their registered e-mail account and to check it for the link

`/verify/:token` is a param route that will send a request to the backend instead of leading the user to a random backend route upon clicking the e-mail link. This way the flow stays inside the UI.

The rest has to do with supporting this change, such as redirecting the user to the `verify-email` route upon sign-up, and adding the Routes to App.jsx.